apps/ecommerce/services/checkout/requirements.txt
ecs-logging==1.0.0
flask==2.0.2
flask-cors==3.0.10
opentelemetry-api==1.9.1
opentelemetry-exporter-otlp==1.9.1
opentelemetry-sdk==1.9.1
opentelemetry-instrumentation-flask==0.28b1
opentelemetry-instrumentation-requests==0.28b1
requests==2.27.1
